requirements.txt
openai>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=2.1.3
openpyxl>=3.1.2
python-dotenv>=1.0.0
django>=5.0
python-decouple
django-widget-tweaks
